# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * IDDET Ads API for an embedded Shopify Dev App
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ShopifyConnectionStatus = typeof ShopifyConnectionStatus[keyof typeof ShopifyConnectionStatus];


export const ShopifyConnectionStatus = {
  connected: 'connected',
  needs_setup: 'needs_setup',
  needs_reauth: 'needs_reauth',
} as const;

export interface ShopifyConnection {
  connected: boolean;
  /** @nullable */
  storeDomain: string | null;
  /** @nullable */
  storeName: string | null;
  status: ShopifyConnectionStatus;
  appUrl: string;
  /** @nullable */
  callbackPath: string | null;
  /** @nullable */
  authUrl: string | null;
  /** @nullable */
  shopifyUserName: string | null;
  iddetConnected: boolean;
}

export type ProductStatus = typeof ProductStatus[keyof typeof ProductStatus];


export const ProductStatus = {
  active: 'active',
  draft: 'draft',
  archived: 'archived',
} as const;

export type ProductAdStatus = typeof ProductAdStatus[keyof typeof ProductAdStatus];


export const ProductAdStatus = {
  not_started: 'not_started',
  draft: 'draft',
  published: 'published',
} as const;

export interface Product {
  id: string;
  title: string;
  description: string;
  price: number;
  currency: string;
  /** @nullable */
  imageUrl: string | null;
  status: ProductStatus;
  adStatus: ProductAdStatus;
  createdAt: string;
}

export interface IddetAccount {
  id: string;
  username: string;
  avatarUrl: string;
  connectedAt: string;
}

export interface IddetConnectInput {
  /** @minLength 3 */
  username: string;
  /** @minLength 6 */
  password: string;
}

export interface IddetSession {
  authenticated: boolean;
  /** @nullable */
  username: string | null;
  /** @nullable */
  avatarUrl: string | null;
  /** @nullable */
  shopDomain: string | null;
}

export interface IddetCommunity {
  id: string;
  slug: string;
  name: string;
  memberCount: number;
  isMember: boolean;
}

export interface AdDraftInput {
  productId: string;
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  content: string;
  accountId: string;
  communityId: string;
}

export type AdDraftStatus = typeof AdDraftStatus[keyof typeof AdDraftStatus];


export const AdDraftStatus = {
  draft: 'draft',
  published: 'published',
} as const;

export interface AdDraft {
  id: string;
  productId: string;
  productTitle: string;
  title: string;
  content: string;
  /** @nullable */
  imageUrl: string | null;
  destinationUrl: string;
  status: AdDraftStatus;
  accountId: string;
  accountUsername: string;
  communityId: string;
  communityName: string;
  /** @nullable */
  publishedAt: string | null;
  createdAt: string;
}

export type ActivityType = typeof ActivityType[keyof typeof ActivityType];


export const ActivityType = {
  product: 'product',
  draft: 'draft',
  publish: 'publish',
  connection: 'connection',
} as const;

export interface Activity {
  id: string;
  type: ActivityType;
  label: string;
  detail: string;
  createdAt: string;
}

export type OverviewStoreStatus = typeof OverviewStoreStatus[keyof typeof OverviewStoreStatus];


export const OverviewStoreStatus = {
  connected: 'connected',
  needs_setup: 'needs_setup',
  needs_reauth: 'needs_reauth',
} as const;

export interface Overview {
  storeStatus: OverviewStoreStatus;
  /** @nullable */
  storeName: string | null;
  productCount: number;
  draftCount: number;
  publishedCount: number;
  connectedAccountCount: number;
  publishedThisWeek: number;
  activity: Activity[];
}

export type StartShopifyOAuthParams = {
/**
 * @pattern ^[a-zA-Z0-9][a-zA-Z0-9-]*\.myshopify\.com$
 */
shop: string;
};

export type CompleteShopifyOAuthParams = {
shop: string;
code: string;
state: string;
hmac: string;
};