# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: IDDET Ads API for an embedded Shopify Dev App
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/iddet/login:
    post:
      operationId: loginWithIddet
      tags: [iddet]
      summary: Sign in with IDDET credentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IddetConnectInput"
      responses:
        "200":
          description: IDDET login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IddetSession"
        "401":
          description: Invalid IDDET credentials
...
components:
  schemas:
...
    IddetConnectInput:
      type: object
      properties:
        username:
          type: string
          minLength: 3
        password:
          type: string
          minLength: 6
          writeOnly: true
      required: [username, password]
    IddetSession:
      type: object
      properties:
        authenticated:
          type: boolean
        username:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        shopDomain:
          type: ["string", "null"]
      required: [authenticated, username, avatarUrl, shopDomain]
...