# lib/api-zod/src/generated/types/iddetSession.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * IDDET Ads API for an embedded Shopify Dev App
 * OpenAPI spec version: 0.1.0
 */

export interface IddetSession {
  authenticated: boolean;
  /** @nullable */
  username: string | null;
  /** @nullable */
  avatarUrl: string | null;
  /** @nullable */
  shopDomain: string | null;
}